+++ CHANGELOG.rst
@@ -4,10 +4,11 @@ What's new
 In development
 ~~~~~~~~~~~~~~
 
-- use AppVeyor for Windows continuous integration
+- use AppVeyor for Windows continuous integration and for .exe freezing
 - fix bugs in tests suite where files were not properly closed
 - add Book.close function to close properly files
 - depend on enum-compat instead of directly enum34
+- add simple script to import/export prices from a gnucash book
 
 Version 0.8.3 (2015-11-01)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~